+++ tik_manager/SmHoudini.py
@@ -625,18 +625,18 @@ class HoudiniManager(RootManager, HoudiniCoreFunctions):
         inputDir = QtWidgets.QFileDialog.getExistingDirectory()
         return os.path.normpath(inputDir)
 
-    def _getTimelineRanges(self):
-        pass
-        R_ast = int(hou.playbar.frameRange()[0])
-        R_min = int(hou.playbar.playbackRange()[0])
-        R_max = int(hou.playbar.playbackRange()[1])
-        R_aet = int(hou.playbar.frameRange()[1])
-        return [R_ast, R_min, R_max, R_aet]
-
-    def _setTimelineRanges(self, rangeList):
-        """Sets the timeline ranges [AnimationStart, Min, Max, AnimationEnd]"""
-        hou.playbar.setFrameRange(rangeList[0], rangeList[3])
-        hou.playbar.setPlaybackRange(rangeList[1], rangeList[2])
+    # def _getTimelineRanges(self):
+    #     pass
+    #     R_ast = int(hou.playbar.frameRange()[0])
+    #     R_min = int(hou.playbar.playbackRange()[0])
+    #     R_max = int(hou.playbar.playbackRange()[1])
+    #     R_aet = int(hou.playbar.frameRange()[1])
+    #     return [R_ast, R_min, R_max, R_aet]
+    #
+    # def _setTimelineRanges(self, rangeList):
+    #     """Sets the timeline ranges [AnimationStart, Min, Max, AnimationEnd]"""
+    #     hou.playbar.setFrameRange(rangeList[0], rangeList[3])
+    #     hou.playbar.setPlaybackRange(rangeList[1], rangeList[2])
 
         #
     def _createCallbacks(self, handler):