+++ ChangeLog.rst
@@ -59,6 +59,10 @@ Tik Manager Version History
 * **Update.606** New Project Naming Conventions added for further customization. All fields defined in the Shared Settings/Conventions/New Project are mandatory
 * **New.700** Project Report and progress
 * **Update.701** Nuke / Loading method changes. Instead of opening a new Nuke instance, it closes the current one (after asking to save first) and opens the new one.
+* **Update.701** Inno Setup default directory for common folder moved under My Documents for possible IO issues
+* **Fix.701** Bug fixed with the Grand Total Work Hours in project report
+* **Update.701** "Create Base Scene To:" feature added to the right click menu in Reference Mode
+* **Update.701** Added Extra Columns options to the Base Scene List
 
 2.5
 ---